refactoring IModelLoader
