+++ packages/modules/viewer-react/src/components/app-ui/providers/index.ts
@@ -2,10 +2,5 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-
-export * from "./useTheme";
-export * from "./useUiProviders";
-export * from "./useIsMounted";
-export * from "./useBaseViewerInitializer";
-export * from "./useAccessToken";
-export * from "./useFrontstages";
+export * from "./DefaultContentGroupProvider";
+export * from "./BackstageItemsProvider";
